+++ src/main/java/net/geforcemods/securitycraft/blockentities/ScannerDoorBlockEntity.java
@@ -68,7 +68,7 @@ public class ScannerDoorBlockEntity extends SpecialDoorBlockEntity implements IV
 					isAllowed = isAllowed(viewingPlayer.getName());
 				}
 				else {
-					viewingPlayer = isOwnedIgnoringMask(player) ? PlayerUtils.getOwnerFromPlayerOrMask(player) : new Owner(player);
+					viewingPlayer = isOwnedBy(player, true) ? PlayerUtils.getOwnerFromPlayerOrMask(player) : new Owner(player);
 					isAllowed = isAllowed(player);
 				}