+++ src/main/java/net/geforcemods/securitycraft/blockentities/ScannerTrapdoorBlockEntity.java
@@ -69,7 +69,7 @@ public class ScannerTrapdoorBlockEntity extends DisguisableBlockEntity implement
 					isAllowed = isAllowed(viewingPlayer.getName());
 				}
 				else {
-					viewingPlayer = isOwnedIgnoringMask(player) ? PlayerUtils.getOwnerFromPlayerOrMask(player) : new Owner(player);
+					viewingPlayer = isOwnedBy(player, true) ? PlayerUtils.getOwnerFromPlayerOrMask(player) : new Owner(player);
 					isAllowed = isAllowed(player);
 				}