+++ Changelog.md
@@ -5,6 +5,8 @@
 - New: In-game notification message that informs players that an error occurred while capturing a frame feed
 - New: Debug feature to help identify other mods preventing players from viewing a Security Camera
 - New: Configuration option for limiting the amount of chunks that may be forceloaded by frame feeds
+- API: New IOwnable#isOwnedBy overload taking a Player and a boolean for checking ownership respecting the new Incognito Mask
+- API: New method Owner#isDefaultOwner to check if the Owner object has no player data associated with it
 - Fix: Several mod incompatibilities that caused frame feeds to not show up correctly
 - Fix: Game performance does not improve when breaking frames with an active frame feed
 - Fix: Clouds sometimes flicker rapidly between two positions when a Frame is displaying a camera feed